gatt: Create hash table to store CCC by client
This commit creates a hash table to store CCC values for each
characteristic that has a CCC descriptor.

The CCC descriptor is implicity created for all characteristics that
have NOTIFY or INDICATE properties, with specific read and write
callbacks that operates over the CCC hash table.

Currently those callbacks are stubs.