gatt: Replace WriteValue by Set Properties on Server
This patch changes the write characteristic value for external GATT
services based on the discussion with the maintainers. DBus Properties
Set for "Value" should be used instead. Offset doesn't need to be
exposed to the API since the value can be cached and internally
bluetoothd can be the bridge between ATT write requests and the
external GATT service.